requirements.txt
Django==4.2.1
pandas==2.0.3
numpy==1.26.4
PyMuPDF==1.23.26
spacy==3.7.4
nltk==3.8.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1.tar.gz
pytesseract==0.3.10
Pillow==10.1.0